# swaggerv2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ecommerce RESTful API",
    "version": "2.0.0",
    "description": "A backend RESTful API for an Ecommerce Store",
    "contact": {
      "name": "Custom Haven",
      "url": "https://github.com/CustomHaven"
    },
    "license": {
      "name": "MIT",
      "url": "https://choosealicense.com/licenses/mit/"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4001/",
      "variables": {}
    },
    {
      "url": "https://rest-api-no-orm.herokuapp.com/",
      "variables": {}
    }
  ],
  "paths": {
...
    "/v2/orders": {
      "summary": "Completed successful orders",
      "description": "All routes to completed orders",
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "All orders made",
        "description": "All orders that were made for the Ecommerce online store",
        "operationId": "all_orders",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "[{\"id\": \"order 1\", \"user_id\": \"user 2\", \"cart_id\": 1, \"status_shipping\": \"success\", \"final_price\": 150.42, \"total_items\": 18, \"tracking_id\": \"some good RM id\", \"created\": \"07-09-2021\", \"modified\": \"15-03-2022\"}, {\"id\": \"order 2\", \"user_id\": \"user 3\", \"cart_id\": 3, \"status_shipping\": \"success\", \"final_price\": 26.74, \"total_items\": 13, \"tracking_id\": \"some good RM id\", \"created\": \"08-09-2021\", \"modified\": \"15-01-2022\"}, {\"id\": \"order 3\", \"user_id\": \"user 3\", \"cart_id\": 4, \"status_shipping\": \"pending\", \"final_price\": 76.73, \"total_items\": 12, \"tracking_id\": \"not available\", \"created\": \"08-08-2022\", \"modified\": \"08-08-2022\"}]"
                  }
                }
              }
            },
            "description": "Success"
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Make a new order",
        "description": "When a customer successfully makes a purchase we save their purchase to a unique table",
        "operationId": "complete_order",
        "requestBody": {
          "description": "Sign up a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\n  \"id\": \"order 4\",\n  \"user_id\": \"user 1\",\n  \"cart_id\": 6,\n  \"status_shipping\": \"pending\",\n  \"final_price\": 21.62,\n  \"total_items\": 4,\n  \"tracking_id\": \"not available\",\n  \"created\": \"20-08-2022\",\n  \"modified\": \"20-08-2022\",\n  \"order_list\": [\n    {\n      \"id\": 11,\n      \"product_id\": \"prod 5\",\n      \"order_id\": \"order 4\",\n      \"quantity\": 2,\n      \"price\": 5.34,\n      \"created\": \"20-08-2022\",\n      \"modified\": \"20-08-2022\"\n    },\n    {\n      \"id\": 12,\n      \"product_id\": \"prod 7\",\n      \"order_id\": \"order 4\",\n      \"quantity\": 2,\n      \"price\": 16.28,\n      \"created\": \"20-08-2022\",\n      \"modified\": \"20-08-2022\"\n    }\n  ]\n}"
                  }
                }
              }
            },
            "description": "Success"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Conflict\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/v2/orders/{user_id}/{cart_id}": {
      "summary": "Completed successful orders",
      "description": "All routes to completed orders",
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Elite route makes a new order and also fills in all order lists",
        "description": "When a customer successfully makes a purchase we save their purchase to a unique table",
        "operationId": "complete_order_does_order_list_as_well",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "The id of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cart_id",
            "in": "path",
            "description": "The id of the cart",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Sign up a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\n  \"id\": \"order 4\",\n  \"user_id\": \"user 1\",\n  \"cart_id\": 6,\n  \"status_shipping\": \"pending\",\n  \"final_price\": 21.62,\n  \"total_items\": 4,\n  \"tracking_id\": \"not available\",\n  \"created\": \"20-08-2022\",\n  \"modified\": \"20-08-2022\",\n  \"order_list\": [\n    {\n      \"id\": 11,\n      \"product_id\": \"prod 5\",\n      \"order_id\": \"order 4\",\n      \"quantity\": 2,\n      \"price\": 5.34,\n      \"created\": \"20-08-2022\",\n      \"modified\": \"20-08-2022\"\n    },\n    {\n      \"id\": 12,\n      \"product_id\": \"prod 7\",\n      \"order_id\": \"order 4\",\n      \"quantity\": 2,\n      \"price\": 16.28,\n      \"created\": \"20-08-2022\",\n      \"modified\": \"20-08-2022\"\n    }\n  ]\n}"
                  }
                }
              }
            },
            "description": "Success"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Conflict\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/v2/orders/{id}": {
      "summary": "Completed successful orders",
      "description": "All routes to completed orders",
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Get an order along with all order items",
        "description": "This route gets and shows an order along with all order items",
        "operationId": "order_with_all_order_items",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\"order\":{ \"id\": \"order 3\", \"user_id\": \"user 2\", \"cart_id\": 4,\n\"status_shipping\": \"pending\", \"final_price\": 76.73, \"total_items\": 12, \"tracking_id\": \"not available\", \"created\": \"08-08-2022\", \"modified\":\n\"08-08-2022\" }, \"order_list\":[{ \"id\": 7,\n\"product_id\": \"prod 7\", \"order_id\": \"order 3\", \"quantity\": 4, \"price\": 32.56, \"created\":\n\"08-08-2022\", \"modified\": \"08-08-2022\" }, { \"id\": 8,\n\"product_id\": \"prod 8\", \"order_id\": \"order 3\", \"quantity\": 3, \"price\": 12.81, \"created\":\n\"08-08-2022\", \"modified\": \"08-08-2022\" }, { \"id\": 9,\n\"product_id\": \"prod 9\", \"order_id\": \"order 3\", \"quantity\": 5, \"price\": 29.35, \"created\":\n\"08-08-2022\", \"modified\": \"08-08-2022\" }, { \"id\": 10,\n\"product_id\": \"prod 10\", \"order_id\": \"order 3\", \"quantity\": 1, \"price\": 2.01, \"created\":\n\"08-08-2022\", \"modified\": \"08-08-2022\" }\n]}"
                  }
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"No items in the order\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "put": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Update an order",
        "description": "Update the tracking and shipping information",
        "operationId": "update_an_orders_shipping_tracking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update the order",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"id\": \"order 3\", \"user_id\": \"user 2\", \"cart_id\": 4,\n\"status_shipping\": \"pending\", \"final_price\": 76.73, \"total_items\": 12, \"tracking_id\": \"not available\", \"created\": \"08-08-2022\", \"modified\":\n\"08-08-2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Forbidden\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/v2/order-list/order/{order_id}": {
      "summary": "The order with all it's items",
      "description": "Current order will display along with all the items that it contains",
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Add item to cart",
        "description": "Add an item to cart based on the cart id",
        "operationId": "add_item_to_the_order",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "description": "The id of the order",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Sign up a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Order-List"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"id\": 11, \"product_id\": \"prod 3\", \"order_id\": \"order 3\", \"quantity\": 3, \"price\": \"24.99\", \"created\": \"04-02-2022\", \"modified\": \"07-04-2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Conflict\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Order List"
        ]
      }
    },
    "/v2/order-list": {
      "summary": "The list of items in a completed order",
      "description": "These paths is all about connecting each items to their particular order",
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Every item in the order",
        "description": "This is all search route only for admins showing every item in each and every order",
        "operationId": "every_item_in_every_order",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "[{ \"id\": 1, \"product_id\": \"prod 1\", \"order_id\": \"order 1\", \"quantity\": 6, \"price\": 46.08, \"created\": \"26-08-2021\", \"modified\": \"02-04-2022\" }, { \"id\": 2, \"product_id\": \"prod 2\", \"order_id\": \"order 1\", \"quantity\": 6, \"price\": 54.36, \"created\": \"08-09-2021\", \"modified\": \"30-05-2022\" }, { \"id\": 3, \"product_id\": \"prod 3\", \"order_id\": \"order 1\", \"quantity\": 6, \"price\": 49.98, \"created\": \"30-10-2021\", \"modified\": \"03-06-2022\" }, { \"id\": 4, \"product_id\": \"prod 4\", \"order_id\": \"order 2\", \"quantity\": 4, \"price\": 8.56, \"created\": \"12-09-2021\", \"modified\": \"18-04-2022\" }, { \"id\": 5, \"product_id\": \"prod 5\", \"order_id\": \"order 2\", \"quantity\": 6, \"price\": 16.02, \"created\": \"12-11-2021\", \"modified\": \"20-03-2022\" }, { \"id\": 6, \"product_id\": \"prod 6\", \"order_id\": \"order 2\", \"quantity\": 3, \"price\": 2.16, \"created\": \"01-09-2021\", \"modified\": \"28-02-2022\" }, { \"id\": 7, \"product_id\": \"prod 7\", \"order_id\": \"order 3\", \"quantity\": 4, \"price\": 32.56, \"created\": \"08-08-2022\", \"modified\": \"08-08-2022\" }, { \"id\": 8, \"product_id\": \"prod 8\", \"order_id\": \"order 3\", \"quantity\": 3, \"price\": 12.81, \"created\": \"08-08-2022\", \"modified\": \"08-08-2022\" }, { \"id\": 9, \"product_id\": \"prod 9\", \"order_id\": \"order 3\", \"quantity\": 5, \"price\": 29.35, \"created\": \"08-08-2022\", \"modified\": \"08-08-2022\" }, { \"id\": 10, \"product_id\": \"prod 10\", \"order_id\": \"order 3\", \"quantity\": 1, \"price\": 2.01, \"created\": \"08-08-2022\", \"modified\": \"08-08-2022\" }]"
                  }
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"No ordered items\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Order List"
        ]
      }
    },
    "/v2/order-list/{id}": {
      "summary": "Route for the specific item in the order",
      "description": "Modify the specific item that is inside cart",
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Get an item in from the order",
        "description": "Retrieve a specific item from the order",
        "operationId": "get_order_item",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the order item",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"id\": 10, \"product_id\": \"prod 10\", \"order_id\": \"order 3\", \"quantity\": 3, \"price\": \"6.03\", \"created\": \"09-01-2022\", \"modified\": \"09-03-2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\"message\": \"Order item not found\"}"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Order List"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Order-List": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "product_id": {
            "type": "string"
          },
          "order_id": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Order-List"
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "cart_id": {
            "type": "number"
          },
          "shipping_status": {
            "type": "string"
          },
          "final_price": {
            "type": "number"
          },
          "total_items": {
            "type": "number"
          },
          "tracking_id": {
            "type": "number"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Order"
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect_sid"
      }
    }
  }
}